Add files via upload
I am trying to lot a graph where the funding ratio of the short bond portfolio should increase with increasing interest rates, the long bond portfolio funding ratio should decrease with increasing interest rates, and a portfolio where funds were allocated to both the short and long bond has a funding ratio this constantly around 1. what corrections should I cake to the code below so as to achieve this. The result should show that the portfolio with parts of both the long and short bonds is the best option for an assured future funding ration of at least 1.